[Sparse]: Optimize code style
